+++ src/contract/index.spec.ts
@@ -1,0 +1,3 @@
+import test from "./test";
+
+describe('contract', test);
