+++ src/contract/wrapper.ts
@@ -1,0 +1,11 @@
+import individualTest from "./individualTest";
+
+export function wrapper(fileName, operations, client) {
+    describe(`${fileName}`, () => {
+        // describe function
+        operations.forEach((operation) => {
+        it(`${operation.iid} @contract`, individualTest(operation, client));
+        });
+    });
+}
+
